+++ fst-chat-back/src/app.module.ts
@@ -4,6 +4,7 @@ import { AppService } from './app.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { UserModule } from './user/module/user.module';
+import { AuthModule } from './auth/module/auth.module';
 
 @Module({
   imports: [
@@ -17,7 +18,7 @@ import { UserModule } from './user/module/user.module';
       }),
       inject: [ConfigService],
     }),
-    UserModule,
+    UserModule,AuthModule
   ],
   controllers: [AppController],
   providers: [AppService],
