+++ fst-chat-back/src/auth/DTO/login-user.dto.ts
@@ -1,4 +1,4 @@
-﻿import { IsEmail, MinLength } from 'class-validator';
+import { IsEmail, MinLength } from 'class-validator';
 
 export class LoginUserDto {
   @IsEmail()
