+++ fst-chat-back/src/auth/DTO/register-user.dto.ts
@@ -1,4 +1,4 @@
-﻿import {
+import {
   IsEmail,
   IsIn,
   IsNotEmpty,
