+++ fst-chat-back/src/user/DTO/UpdateUserDTO.ts
@@ -1,5 +1,5 @@
 import { OmitType } from '@nestjs/swagger';
-import { CreateUserDTO } from './CreateUserDTO';
+import { CreateUserDTO } from '../../auth/DTO/CreateUserDTO';
 import { IsMongoId } from 'class-validator';
 
 export class UpdateUserDTO extends OmitType(CreateUserDTO, ['email'] as const) {
